requirements.txt
streamlit>=1.32,<2
opencv-python-headless==4.12.0.88
Pillow>=10,<12
numpy>=1.26,<2.1
